Both Maps Were Added
Both of the maps were added to the game and I fixed the issues with 1 of the maps and tried my best to fix the other map. Changes need to be made in the way that maps are made so that they are not as buggy as they currently are.