Fixed the AI and other changes
The AI is fixed, so that it now does not get bugged when shooting at the player. It also now has a idle state, and will not get bugged within terrain when patrolling because it will draw a navmeshpath before moving to the desired point.

There are also other changes like the movement speed and changes to air strafing to make the player more agile in the air.